modal component debugging

+++ database/migrations/2024_05_22_130802_add_columns_qlik_users_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::table('qlik_users', function (Blueprint $table) {
             //
-            $table->unsignedInteger('qlik_conf_id');
+            $table->unsignedBigInteger('qlik_conf_id');
             $table->foreign('qlik_conf_id')->references('id')->on('qlik_confs');
         });
     }
